\name{NGS.Tools.BICseq-package}
\alias{NGS.Tools.BICseq-package}
\alias{NGS.Tools.BICseq}
\docType{package}
\title{
BIC-seq copy number analysis pipeline.
}
\description{
Generate a standard data CNV analysis using BIC-seq as the CNV calling algorithm.
}
\details{
\tabular{ll}{
Package: \tab NGS.Tools.BICseq\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2015-08-21\cr
License: \tab GPL-2\cr
}
get.bicseq.chr run.bicseq.cnv.pipeline
}
\author{
Richard de Borja

Maintainer: Richard de Borja <richard.deborja@sickkids.ca>
}
\references{
http://compbio.med.harvard.edu/Supplements/PNAS11.html
}
\keyword{ BICseq }
\keyword{ copy-number }
\keyword{ CNV }
